refactor: :fire: remove unefficent graph
